Make TestTargetMetadataProvider top-level class
